+++ lib/index.d.ts
@@ -29,7 +29,12 @@ export declare type xarrowPropsType = {
     divContainerProps?: React.HTMLProps<HTMLDivElement>;
     SVGcanvasStyle?: React.CSSProperties;
     divContainerStyle?: React.CSSProperties;
-    extendSVGcanvas?: number;
+    _extendSVGcanvas?: number;
+    _debug: boolean;
+    _cpx1Offset: number;
+    _cpy1Offset: number;
+    _cpx2Offset: number;
+    _cpy2Offset: number;
 };
 export declare type anchorType = anchorPositionType | anchorCustomPositionType;
 export declare type anchorPositionType = "middle" | "left" | "right" | "top" | "bottom" | "auto";